+++ src/main/java/adapter/PaybackCard.java
@@ -2,10 +2,9 @@ package adapter;
 
 public class PaybackCard {
 
-    private static final int AVAILABLE = 500;
+    private static final float AVAILABLE = 500;
 
-
-    public boolean payByPoints(Float amount) {
+    public boolean payByPoints(float amount) {
         return amount <= AVAILABLE;
     }
